# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API demo with auth",
        "title": "go-demo API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/sql-logs": {
            "get": {
                "description": "Provide database name via query parameter \"db\" to list its SQL queries.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sql-logs"
                ],
                "summary": "List SQL queries by database",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Database name",
                        "name": "db",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ListByDBResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/v1/sql-logs/databases": {
            "get": {
                "description": "Returns distinct database names that have SQL log entries.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sql-logs"
                ],
                "summary": "List databases with SQL logs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ListDatabasesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string"
                        },
                        "message": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "handlers.ListByDBResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.SQLLogItem"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handlers.ListDatabasesResponse": {
            "type": "object",
            "properties": {
                "databases": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "handlers.SQLLogItem": {
            "type": "object",
            "properties": {
                "exec_count": {
                    "type": "integer"
                },
                "exec_time_ms": {
                    "type": "integer"
                },
                "sql_query": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}